Put common fixtures in conftest
