+++ tests/make_100_tsv_test.py
@@ -15,16 +15,6 @@ def tsv_file():
 
 
 @pytest.fixture(scope="module")
-def temp_dir():
-    return Path(mkdtemp())
-
-
-@pytest.fixture(scope="module")
-def db_path(temp_dir):
-    return temp_dir / "sca.sqlite3"
-
-
-@pytest.fixture(scope="module")
 def sca_filled(db_path, tsv_file):
     corpus = sca.from_file(
         db_path=db_path,
